move scoring trace into validator/scoring_trace.py
Pulls log_scoring_trace / non_earner_lines / diagnose_non_earner out of
scoring.py so the latter stays focused on the replay algorithm
(483 -> 375 lines). The new module is pure presentation — it never
mutates state.

Precompute the per-hotkey success_rate map in calculate_miner_rewards
and pass it to log_scoring_trace; this kills the reverse import from
scoring_trace -> scoring (no more circular feel) and removes redundant
success_rate calls in the reward loop. DirectionTrace stays in
scoring.py since that's where replay_crown_time_window mutates it; the
trace module references it via TYPE_CHECKING only.

utils/logging.py was deliberately not chosen: it's generic log infra
(rotating handlers, EVENT level) and these helpers are scoring-specific
domain logic — wrong layer.
